backend/requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0